Generate docs corpus at build time for filesystem-less hosts
The chat route previously read content/docs from disk at request time,
which fails on hosts without a runtime filesystem (e.g. Cloudflare
Workers). The corpus is now generated by scripts/generate-docs-corpus.mjs
during postinstall and build, and bundled as a JSON module import, so the
route works the same on any host.
