+++ package.json
@@ -4,10 +4,10 @@
   "private": true,
   "packageManager": "pnpm@10.30.3",
   "scripts": {
-    "build": "next build",
+    "build": "node scripts/generate-docs-corpus.mjs && next build",
     "dev": "next dev",
     "start": "next start",
-    "postinstall": "fumadocs-mdx",
+    "postinstall": "fumadocs-mdx && node scripts/generate-docs-corpus.mjs",
     "watch": "fumadocs-mdx watch"
   },
   "dependencies": {
